Disable Due/Completed column from appearing in worktray
